+++ packages/benchmark-js/src/types.internal.ts
@@ -7,6 +7,7 @@ export type _Nanosecond = number & { __ns__: void };
 export interface _BenchmarkSettings {
     delay?: _Nanosecond;
     initCount?: number;
+    maxAdjustTime?: _Nanosecond;
     maxPreparingTime?: _Nanosecond;
     maxTime?: _Nanosecond;
     minSamples?: number;